requirements.txt
annotated-types==0.7.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
google-ai-generativelanguage==0.6.10
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
grpcio==1.67.1
grpcio-status==1.67.1
gTTS==2.5.3
httplib2==0.22.0
idna==3.10
joblib==1.4.2
nltk==3.9.1
playsound==1.3.0
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
pyparsing==3.2.0
python-dotenv==1.0.1
pyttsx3==2.98
regex==2024.9.11
requests==2.32.3
rsa==4.9
simpleaudio==1.0.4
sounddevice==0.5.1
SpeechRecognition==3.11.0
textblob==0.18.0.post0
tqdm==4.66.6
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
